feat(CreateAndUpdateProduct): Create server actions and components to create and update the products | :sparkles:
